+++ app/chains/user/user.routes.js
@@ -5,6 +5,7 @@
  */
 var userController = require('./user.controller');
 var responseHandler = require('../../utils/response.handler');
+var authenticationHandler = require('../../utils/authentication.handler');
 var authentication = require('../../middleware/authentication.middleware');
 
 module.exports = function(routes) {
@@ -26,7 +27,7 @@ module.exports = function(routes) {
     //get user by id
     routes.get('/:id', function(request, response) {
         userController.getUserById(request.params.id)
-            .then(responseHandler.trimByAttributes(request.headers['x-forwarded-email'], ['canViewProfile', 'canEditProfile'], response))
+            .then(authenticationHandler.trimByAttributes(request.headers['x-forwarded-email'], ['canViewProfile', 'canEditProfile']))
             .then(responseHandler.sendJsonResponse(response))
             .catch(responseHandler.sendErrorResponse(response));
