+++ app/utils/response.handler.js
@@ -162,109 +162,3 @@ exports.sendThumbnailResponse = function(response) {
         response.download(config.UPLOAD_PATH + generatedName);
     };
 };
-
-// TRIMMING
-// ==============================================================================
-// TODO: error handling
-
-exports.trimByAttributes = function(email, attributes, response) {
-    return function(body) {
-        return q.promise(function(resolve, reject) {
-            var userIsSelf = authenticationHandler.isSelf(email, body._id);
-
-            return q.all(userIsSelf)
-                .then(function(result) {
-                    if (result) {
-                        return resolve(body);
-                    } else {
-                        return authenticationHandler.getUserAttributeObjects(email)
-                            .then(extractRelevantAttributes(attributes))
-                            .then(findSmallestIntersectionOfHiddenFields)
-                            .then(trimByFields(body))
-                            .then(resolve)
-                            .catch(exports.sendUnauthorizedResponse(response));
-                    }
-                });
-        });
-    };
-};
-
-function extractRelevantAttributes(relevantAttributes) {
-    return function(presentAttributes) {
-        return q.promise(function(resolve, reject) {
-            if (presentAttributes.length <= 0) {
-                reject();
-            }
-
-            var list = [];
-            presentAttributes.forEach(function(presentAttribute) {
-                if (relevantAttributes.indexOf(presentAttribute.name) >= 0) {
-                    list.push(presentAttribute);
-                }
-            });
-
-            return q.all(list)
-                .then(resolve);
-        });
-    };
-}
-
-function findSmallestIntersectionOfHiddenFields(attributes) {
-    return q.promise(function(resolve) {
-        var list = [];
-        var length = attributes.length;
-
-        return getFieldCounts(attributes)
-            .then(pushFieldsIfPresentInAllAttributes(length))
-            .then(resolve);
-    });
-}
-
-function getFieldCounts(attributes) {
-    return q.promise(function(resolve) {
-        var count = {};
-        attributes.forEach(function(attribute) {
-            if (attribute.hiddenFields) {
-                attribute.hiddenFields.forEach(function(hiddenField) {
-                    if (!count[hiddenField]) {
-                        count[hiddenField] = 1;
-                    } else {
-                        count[hiddenField]++;
-                    }
-                });
-            }
-        });
-
-        return resolve(count);
-    });
-}
-
-function pushFieldsIfPresentInAllAttributes(length) {
-    return function(count) {
-        return q.promise(function(resolve) {
-            var promises = [];
-            for (var field in count) {
-                if (count[field] === length) {
-                    promises.push(field);
-                }
-            }
-
-            q.all(promises)
-                .then(resolve);
-        });
-    };
-}
-
-function trimByFields(body) {
-    return function(fields) {
-        return q.promise(function(resolve) {
-            for (var field in body) {
-                if (fields.indexOf(field) >= 0) {
-                    body[field] = null;
-                }
-            }
-
-            return resolve(body);
-        });
-    };
-}
